Major refactoring of the task code
Previous task code would not allow irregular task blocks,
and a result would have a problematic dychotomy of format between
repeated task block and concatenated records of different task
blocks.

This refactoring fixes this, but in the process must abandon the
numpy format for saving results, as it does not easily supports
arrays of irregular lengths (not without using pickle anyway, and
if using pickle, then using pickle directly is more interesting;
it allows to add metadata---such as provenance data--- into the
result file directly).

The task code is simplified.